+++ test/build-standalone-notes.js
@@ -131,6 +131,9 @@ const testNotesHostBridge = `
         postDailyNotes: function(type, dayOffset, currentDate) {
             window.__testApi.messages.push({ type: 'postDailyNotes', subType: type, dayOffset: dayOffset, currentDate: currentDate });
         },
+        importMdFiles: function(filePaths, targetNodeId, position) {
+            window.__testApi.messages.push({ type: 'importMdFiles', filePaths: filePaths, targetNodeId: targetNodeId, position: position });
+        },
         showConfirm: function(id, message) {
             window.__testApi.messages.push({ type: 'showConfirm', id: id, message: message });
         },